move FilterTabGroup into generic components
