+++ res/css/components/views/elements/_FilterTabGroup.pcss
